Make Buf.send() faster, again, for confined buffers
Motivation:
As has been mentioned before, this is likely to be an important operation.

Modification:
Use a dedicated transfer thread to facilitate the handoff of thread-confined memory segments, from one thread to another.
By using a separate thread for this, we no longer need to create and close a shared segment, sparing us of the thread-local handshake costs related to that.

Result:
Sending a thread-confined buffer from one thread to another is now about twice as fast.